Reorder commands, improve help, allow enum in parameters
