+++ src/NeoSharp.Application/Client/Prompt[Invokes].cs
@@ -1,0 +1,38 @@
+﻿using NeoSharp.Application.Attributes;
+using NeoSharp.Core.Models;
+using NeoSharp.Core.Types;
+using System;
+
+namespace NeoSharp.Application.Client
+{
+    public partial class Prompt : IPrompt
+    {
+        /// <summary>
+        /// Invoke contract
+        /// </summary>
+        /// <param name="contractHash">Contract</param>
+        /// <param name="body">Body</param>
+        [PromptCommand("invoke", Help = "Invoke a contract", Category = "Invokes")]
+        private void Invoke(UInt160 contractHash, [PromptCommandParameterBody] object[] args)
+        {
+            Contract contract = Contract.GetContract(contractHash);
+            if (contract == null) throw (new ArgumentNullException("Contract not found"));
+
+            var tx = contract.CreateInvokeTransaction(args);
+        }
+
+        /// <summary>
+        /// Invoke contract
+        /// </summary>
+        /// <param name="contractHash">Contract</param>
+        /// <param name="body">Body</param>
+        [PromptCommand("testinvoke", Help = "Test invoke contract", Category = "Invokes")]
+        private void TestInvoke(UInt160 contractHash, [PromptCommandParameterBody] object[] args)
+        {
+            Contract contract = Contract.GetContract(contractHash);
+            if (contract == null) throw (new ArgumentNullException("Contract not found"));
+
+            var tx = contract.CreateInvokeTransaction(args);
+        }
+    }
+}
